Move isAllowedToAccessInstanceMember
