+++ server/src/services/completion.ts
@@ -10,10 +10,10 @@ import {
     collectParentScopeList,
     isAnonymousIdentifier, SymbolScope
 } from "../compiler_analyzer/symbolScope";
-import {isAllowedToAccessInstanceMember} from "../compiler_analyzer/typeCast";
 import {ComplementHint, ComplementKind} from "../compiler_analyzer/complementHint";
 import {findScopeContainingPosition} from "./serviceHelper";
 import {TextPosition} from "../compiler_tokenizer/textLocation";
+import {isAllowedToAccessInstanceMember} from "../compiler_analyzer/symbolUtils";
 
 /**
  * Returns the completion candidates for the specified position.
